Refactor HttpClient configuration
Configure default HttpClient instance.
Fix incorrect type being used for User Agent version.
Refactor User Agent to include the Git SHA.
Move HttpClient extension methods to their own class.
Set default HttpClient timeout to 20s instead of 100s.
